+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 0.8.10
+
+* Fixed Carousel window width issue
+
 ### 0.8.9
 
 * Fixed overflow on `contained` Carousel